// src/models/IoK8sApiextensionsApiserverPkgApisApiextensionsV1CustomResourceDefinitionSpec/types.ts
import { IoK8sApiextensionsApiserverPkgApisApiextensionsV1CustomResourceConversionModelOptions as IoK8sApiextensionsApiserverPkgApisApiextensionsV1CustomResourceConversion } from '../IoK8sApiextensionsApiserverPkgApisApiextensionsV1CustomResourceConversion';
import { IoK8sApiextensionsApiserverPkgApisApiextensionsV1CustomResourceDefinitionNamesModelOptions as IoK8sApiextensionsApiserverPkgApisApiextensionsV1CustomResourceDefinitionNames } from '../IoK8sApiextensionsApiserverPkgApisApiextensionsV1CustomResourceDefinitionNames';
import { IoK8sApiextensionsApiserverPkgApisApiextensionsV1CustomResourceDefinitionVersionModelOptions as IoK8sApiextensionsApiserverPkgApisApiextensionsV1CustomResourceDefinitionVersion } from '../IoK8sApiextensionsApiserverPkgApisApiextensionsV1CustomResourceDefinitionVersion';

export interface IoK8sApiextensionsApiserverPkgApisApiextensionsV1CustomResourceDefinitionSpecModelOptions {
  conversion?: IoK8sApiextensionsApiserverPkgApisApiextensionsV1CustomResourceConversion;
  group: string;
  names: IoK8sApiextensionsApiserverPkgApisApiextensionsV1CustomResourceDefinitionNames;
  preserveUnknownFields?: boolean;
  scope: string;
  versions: Array<IoK8sApiextensionsApiserverPkgApisApiextensionsV1CustomResourceDefinitionVersion>;
}